requirements.txt
googletrans-py==4.0.0